Following Tutorial Videos Part 2 Removed Phantomail Bundle - see below Removed Stamped Phantomail Bundle - see below Phantomail Stamps trigger callback on right click Phantomailbox updated to BlockEntity and displays correctly ingame Phantomailbox item sprite working correctly Recipes updated for Phantomailbox and Phantomail Stamp Phantomailbox inventory implemented but not yet accessible by player
The process of stamping a Bundle to prepare it for delivery introduced several fundamental problems:
1. Phantomail Bundles were a flat upgrade from normal Bundles
2. The process of applying a stamp to a bundle was cumbersome
3. Bundles arrive in 1.21.2 and the NeoForge version I am learning with is 1.21, so a crafting recipe relying on Bundles would be a problem

The mod workflow has been streamlined to allow a player to insert one Phantomail Stamp and one outgoing item into a Phantomailbox for pickup at a time. A Phantomailbox still has four slots for receiving incoming mail.

The original purpose of the Phantomail mod was to create a use case for the piles of extra Name Tags that players were accumulating on our modded server. At the time, the Name Tag item retained some semblance of being rare or special because Name Tags weren't available as trades from Villagers yet, and our modded server didn't have any mods that added crafting recipes for Name Tags. However, it is now trivial to obtain infinite Name Tags in vanilla survival gameplay by simply establishing a Villager trading farm (or building an auto-fisher), and crafting recipes for Name Tags have become a staple in many popular modpacks, broadly disincentivizing the hoarding of Name Tags.

Sending items long distances quickly is a very powerful utility, and the limited availability of Name Tags seemed like a good balancing factor in the past, but this is definitely no longer the case.

For now, the recipe to craft a Phantomail Stamp will continue to use a Name Tag. Once the mod's primary functions of sending and receiving mail are functional, more attention can be focused on identifying a suitable replacement for Name Tags which is not trivially infinitely obtainable by players.