Progress on item delivery Added code to simulate item pickup or dropoff from courier Items are moved from mailbox to savedata Items are moved from savedata to mailbox
Started working on a primitive slot system - copying and pasting this code is super ugly but I'm not sure how else to do this in Java. I don't understand enough about what's going on with these data structures yet to know when I'm copying them or modifying them with certainty, so a lot of this is YOLO code.

I'm not sure if I'm allowed to just dump these things into some kind of list or not, but for purposes of testing and getting the system working at all, this is functional enough.

With that said, this code is getting messy and silly because it is trying to compress multiple stages into single interactions, so the next step forward is adding the actual courier mob and getting that working.